requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
pydantic>=2.0.0
discord.py>=2.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
